backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4

torch
torchvision
pillow==10.4.0
numpy==2.0.1